Enhanced Swift support by attributing interface with _Nullable & _Nonnull and added a isTouchIDAvailable property
The new Objective-C keywords _Nullable & _Nonnull remove the guessing that the Swift Objective-C importer does when trying to determine if a particular Objective-C pointer will be null or not.

The existing touchIDAvailable function does not import well into Swift as Swift removes the BOOL return and simply turns the method into a 'throws' method.  Thus, in Swift, in order to determine if Touch ID is available, the call needed to be wrapped in a  do {} catch {} block, which is a bit cumbersome.  By providing the isTouchIDAvailable property, Swift code can trivially ask the intended question.